+++ sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/StreamExecution.scala
@@ -17,8 +17,6 @@
 
 package org.apache.spark.sql.execution.streaming
 
-import java.io.{InputStreamReader, OutputStreamWriter}
-import java.nio.charset.StandardCharsets
 import java.util.UUID
 import java.util.concurrent.{CountDownLatch, TimeUnit}
 import java.util.concurrent.locks.ReentrantLock
@@ -26,11 +24,7 @@ import java.util.concurrent.locks.ReentrantLock
 import scala.collection.mutable.ArrayBuffer
 import scala.util.control.NonFatal
 
-import org.apache.commons.io.IOUtils
-import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.fs.{FileSystem, FSDataInputStream, FSDataOutputStream, Path}
-import org.json4s.NoTypeHints
-import org.json4s.jackson.Serialization
+import org.apache.hadoop.fs.Path
 
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql._
@@ -627,63 +621,6 @@ class StreamExecution(
   case object INITIALIZED extends State
   case object ACTIVE extends State
   case object TERMINATED extends State
-}
-
-
-/**
- * Contains metadata associated with a [[StreamingQuery]]. This information is written
- * in the checkpoint location the first time a query is started and recovered every time the query
- * is restarted.
- *
- * @param id  unique id of the [[StreamingQuery]] that needs to be persisted across restarts
- */
-case class StreamMetadata(id: String) {
-  def json: String = Serialization.write(this)(StreamMetadata.format)
-}
-
-object StreamMetadata extends Logging {
-  implicit val format = Serialization.formats(NoTypeHints)
-
-  /** Read the metadata from file if it exists */
-  def read(metadataFile: Path, hadoopConf: Configuration): Option[StreamMetadata] = {
-    val fs = FileSystem.get(hadoopConf)
-    if (fs.exists(metadataFile)) {
-      var input: FSDataInputStream = null
-      try {
-        input = fs.open(metadataFile)
-        val reader = new InputStreamReader(input, StandardCharsets.UTF_8)
-        val metadata = Serialization.read[StreamMetadata](reader)
-        Some(metadata)
-      } catch {
-        case NonFatal(e) =>
-          logError(s"Error reading stream metadata from $metadataFile", e)
-          throw e
-      } finally {
-        IOUtils.closeQuietly(input)
-      }
-    } else None
-  }
-
-  /** Write metadata to file */
-  def write(
-      metadata: StreamMetadata,
-      metadataFile: Path,
-      hadoopConf: Configuration): Unit = {
-    var output: FSDataOutputStream = null
-    try {
-      val fs = FileSystem.get(hadoopConf)
-      output = fs.create(metadataFile)
-      val writer = new OutputStreamWriter(output)
-      Serialization.write(metadata, writer)
-      writer.close()
-    } catch {
-      case NonFatal(e) =>
-        logError(s"Error writing stream metadata $metadata to $metadataFile", e)
-        throw e
-    } finally {
-      IOUtils.closeQuietly(output)
-    }
-  }
 }
